Renamed and updated controller and presenter for editmessage
